backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.3.13
langchain-google-genai==2.0.4
langgraph==0.2.45
langchain-community==0.3.7
langchain-core>=0.3.59,<1.0.0
chromadb==0.5.20
openai==1.76.2
anthropic>=0.51.0,<1.0.0
google-generativeai==0.8.3
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
aiohttp==3.10.11
PyPDF2==3.0.1
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0

pandas==2.0.0
openpyxl==3.1.2
xlrd==2.0.1
tabulate==0.9.0
aiofiles==24.1.0
python-json-logger==2.0.7
pymongo==4.10.1
sqlalchemy>=1.4,<2.0.36
psycopg2-binary==2.9.10

